+++ Dockerfile
@@ -5,7 +5,7 @@ RUN conda install -c conda-forge rdkit -y
 
 COPY ./requirements.txt ./requirements.txt
 
-RUN pip install -r requirements.txt
+RUN pip install --no-cache-dir -r requirements.txt
 
 COPY . .
