+++ requirements.txt
@@ -1,26 +1,40 @@
 alembic==1.13.2
+amqp==5.2.0
 annotated-types==0.7.0
 anyio==4.4.0
 async-timeout==4.0.3
+billiard==4.2.0
+celery==5.4.0
 certifi==2024.7.4
 click==8.1.7
+click-didyoumean==0.3.1
+click-plugins==1.1.1
+click-repl==0.3.0
+colour==0.1.4
 fastapi==0.112.0
 greenlet==3.0.3
 h11==0.14.0
 httpcore==1.0.5
 httpx==0.27.0
 idna==3.7
+infinity==1.5
 iniconfig==2.0.0
+intervals==0.9.2
+kombu==5.4.1
 Mako==1.3.5
 MarkupSafe==2.1.5
 numpy==2.1.0
+orderedmultidict==1.0.1
 packaging==24.1
+passlib==1.7.4
 pillow==10.4.0
 pluggy==1.5.0
+prompt_toolkit==3.0.47
 psycopg2-binary==2.9.9
 pydantic==2.8.2
 pydantic_core==2.20.1
 pytest==8.3.2
+pytest-asyncio==0.24.0
 python-dotenv==1.0.1
 rdkit==2024.3.5
 redis==5.0.8
@@ -28,7 +42,8 @@ setuptools==72.1.0
 sniffio==1.3.1
 SQLAlchemy==2.0.32
 starlette==0.37.2
-typing_extensions==4.12.2
+tzdata==2024.1
 uvicorn==0.30.6
+vine==5.1.0
+wcwidth==0.2.13
 wheel==0.43.0
-sqlalchemy-utils~=0.41.2
