+++ src/models.py
@@ -1,0 +1,10 @@
+from sqlalchemy.orm import declarative_base
+from sqlalchemy import Column, Integer, String
+
+
+Base = declarative_base()
+
+class Molecule(Base):
+    __tablename__ = 'molecules'
+    id = Column(Integer, primary_key=True, index=True)
+    smiles = Column(String, index=True)
