+++ tests/test_main.py
@@ -1,5 +1,4 @@
 from unittest.mock import patch
-
 import pytest
 from fastapi.testclient import TestClient
 from sqlalchemy import create_engine, StaticPool
